refactor(client/java): consolidate test cases
